Extract sync delta observability
