+++ service-worker.js
@@ -147,6 +147,7 @@ const APP_SHELL = [
   '/js/sync-schema.js',
   '/js/sync-delta.js',
   '/js/sync-delta-registry.js',
+  '/js/sync-delta-observability.js',
   '/js/sync-tombstones.js',
   '/js/sync-messenger.js',
   '/js/sync-environment.js',
